Rename core components folder and add test cases
