+++ react-frontend/src/core-components/ImageComponent.test.js
@@ -1,0 +1,14 @@
+import ImageComponent from "./ImageComponent";
+import renderWithRouter from "helpers/test-helpers/routeHelper";
+import { IMG_BASE_PATH } from "constants/appConstants";
+
+describe("Image component test", () => {
+  test("renders image component with image", () => {
+    const { getByAltText } = renderWithRouter(ImageComponent, {
+      alt: "test image",
+      src: `${IMG_BASE_PATH}/high-five.png`,
+    });
+    const testImage = getByAltText("test image");
+    expect(testImage).toBeInTheDocument();
+  });
+});
